Make motor controller wrapper configs a Utils method
- subsystem independent methods
