+++ src/main/java/frc/team3130/robot/subsystems/Flywheel.java
@@ -11,6 +11,8 @@ import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj2.command.Subsystem;
 import frc.team3130.robot.RobotMap;
 
+import static frc.team3130.robot.util.Utils.configPIDF;
+
 public class Flywheel implements Subsystem {
 
 
@@ -177,11 +179,5 @@ public class Flywheel implements Subsystem {
         SmartDashboard.putBoolean("Flywheel canShoot", canShoot());
     }
 
-    public static void configPIDF(WPI_TalonFX _talon, double kP, double kI, double kD, double kF) {
-        _talon.config_kP(0, kP, 0);
-        _talon.config_kI(0, kI, 0);
-        _talon.config_kD(0, kD, 0);
-        _talon.config_kF(0, kF, 0);
-    }
 }
